+++ src/midgard/logging.cc
@@ -1,12 +1,10 @@
 #include "midgard/logging.h"
 
 #include <chrono>
-#include <cstdlib>
 #include <ctime>
 #include <fstream>
 #include <iostream>
 #include <memory>
-#include <sstream>
 #include <stdexcept>
 
 #ifdef __ANDROID__
@@ -14,6 +12,60 @@
 #endif
 
 namespace {
+
+// Highlight special tags according to their levels.
+const std::string& highlight(const std::string& value) {
+  const static std::unordered_map<std::string, std::string>
+      highlighted{{"[ERROR]", "\x1b[31;1m[ERROR]\x1b[0m"},
+                  {"[WARN]", "\x1b[33;1m[WARN]\x1b[0m"},
+                  {"[INFO]", "\x1b[32;1m[INFO]\x1b[0m"},
+                  {"[DEBUG]", "\x1b[34;1m[DEBUG]\x1b[0m"},
+                  {"[TRACE]", "\x1b[37;1m[TRACE]\x1b[0m"}};
+
+  auto it = highlighted.find(value);
+  return (it == highlighted.end()) ? value : it->second;
+}
+
+// Some compilers don't support enum hash specialization, so provide our own one.
+struct EnumHash {
+  template <typename Type> auto operator()(const Type& value) const {
+    using UnderlyingType = std::underlying_type_t<Type>;
+    return std::hash<UnderlyingType>()(static_cast<UnderlyingType>(value));
+  }
+};
+
+// LogLevel to string map.
+// Is used to save backward compatibility with previous logger version.
+const std::unordered_map<valhalla::midgard::logging::LogLevel, std::string, EnumHash>
+    level_tags{{valhalla::midgard::logging::LogLevel::ERROR, "ERROR"},
+               {valhalla::midgard::logging::LogLevel::WARN, "WARN"},
+               {valhalla::midgard::logging::LogLevel::INFO, "INFO"},
+               {valhalla::midgard::logging::LogLevel::DEBUG, "DEBUG"},
+               {valhalla::midgard::logging::LogLevel::TRACE, "TRACE"}};
+
+#ifdef __ANDROID__
+// Determine android log priority using message tags.
+// Only first level tag is used to determine priority. All other are ignored.
+android_LogPriority android_priority(const Tags& tags) {
+  static const std::unordered_map<std::string, android_LogPriority>
+      android_levels{{"ERROR", ANDROID_LOG_ERROR},
+                     {"WARN", ANDROID_LOG_WARN},
+                     {"INFO", ANDROID_LOG_INFO},
+                     {"DEBUG", ANDROID_LOG_DEBUG},
+                     {"TRACE", ANDROID_LOG_VERBOSE}};
+
+  // Get first appropriate tag and return its log priority.
+  for (const auto& tag : tags.tags()) {
+    if (tag.special()) {
+      auto it = android_levels.find();
+      if (it != android_levels.end())
+        return it->second;
+    }
+  }
+
+  return ANDROID_LOG_INFO;
+}
+#endif
 
 inline std::tm* get_gmtime(const std::time_t* time, std::tm* tm) {
 #ifdef _MSC_VER
@@ -43,34 +95,6 @@ std::string TimeStamp() {
           gmt.tm_mday, gmt.tm_hour, gmt.tm_min, fractional_seconds.count());
   return buffer;
 }
-
-// the Log levels we support
-struct EnumHasher {
-  template <typename T> std::size_t operator()(T t) const {
-    return static_cast<std::size_t>(t);
-  }
-};
-const std::unordered_map<valhalla::midgard::logging::LogLevel, std::string, EnumHasher>
-    uncolored{{valhalla::midgard::logging::LogLevel::ERROR, " [ERROR] "},
-              {valhalla::midgard::logging::LogLevel::WARN, " [WARN] "},
-              {valhalla::midgard::logging::LogLevel::INFO, " [INFO] "},
-              {valhalla::midgard::logging::LogLevel::DEBUG, " [DEBUG] "},
-              {valhalla::midgard::logging::LogLevel::TRACE, " [TRACE] "}};
-const std::unordered_map<valhalla::midgard::logging::LogLevel, std::string, EnumHasher>
-    colored{{valhalla::midgard::logging::LogLevel::ERROR, " \x1b[31;1m[ERROR]\x1b[0m "},
-            {valhalla::midgard::logging::LogLevel::WARN, " \x1b[33;1m[WARN]\x1b[0m "},
-            {valhalla::midgard::logging::LogLevel::INFO, " \x1b[32;1m[INFO]\x1b[0m "},
-            {valhalla::midgard::logging::LogLevel::DEBUG, " \x1b[34;1m[DEBUG]\x1b[0m "},
-            {valhalla::midgard::logging::LogLevel::TRACE, " \x1b[37;1m[TRACE]\x1b[0m "}};
-#ifdef __ANDROID__
-const std::unordered_map<valhalla::midgard::logging::LogLevel, android_LogPriority, EnumHasher>
-    android_levels{{valhalla::midgard::logging::LogLevel::ERROR, ANDROID_LOG_ERROR},
-                   {valhalla::midgard::logging::LogLevel::WARN, ANDROID_LOG_WARN},
-                   {valhalla::midgard::logging::LogLevel::INFO, ANDROID_LOG_INFO},
-                   {valhalla::midgard::logging::LogLevel::DEBUG, ANDROID_LOG_DEBUG},
-                   {valhalla::midgard::logging::LogLevel::TRACE, ANDROID_LOG_VERBOSE}};
-#endif
-
 } // namespace
 
 namespace valhalla {
@@ -78,209 +102,137 @@ namespace midgard {
 
 namespace logging {
 
-// a factory that can create loggers (that derive from 'logger') via function pointers
-// this way you could make your own logger that sends log messages to who knows where
-Logger* LoggerFactory::Produce(const LoggingConfig& config) const {
-  // grab the type
-  auto type = config.find("type");
-  if (type == config.end()) {
-    throw std::runtime_error("Logging factory configuration requires a type of logger");
-  }
-  // grab the logger
-  auto found = find(type->second);
-  if (found != end()) {
-    return found->second(config);
-  }
-  // couldn't get a logger
-  throw std::runtime_error("Couldn't produce logger for type: " + type->second);
+const std::string& LevelTag(valhalla::midgard::logging::LogLevel level) {
+  return level_tags.find(level)->second;
 }
 
-// statically get a factory
-LoggerFactory& GetFactory() {
-  static LoggerFactory factory_singleton{};
-  return factory_singleton;
+Tags operator+(Tags lhs, const Tags& rhs) {
+  return lhs += rhs;
 }
 
-// register your custom loggers here
-bool RegisterLogger(const std::string& name, LoggerCreator function_ptr) {
-  auto success = GetFactory().emplace(name, function_ptr);
-  return success.second;
+void Logger::log(std::unique_ptr<Message>&& message) {
+  // Processor can be reconfigured during message enqueuing, so protect it.
+  std::lock_guard<std::mutex> lock(processor_mutex_);
+  processor_->push(std::move(message));
 }
 
-// logger base class, not pure virtual so you can use as a null logger if you want
-Logger::Logger(const LoggingConfig& config){};
-Logger::~Logger(){};
-void Logger::Log(const std::string&, const LogLevel){};
-void Logger::Log(const std::string&, const std::string&){};
-bool logger_registered = RegisterLogger("", [](const LoggingConfig& config) {
-  Logger* l = new Logger(config);
-  return l;
-});
+// Create default tag with its current timestamp.
+Tags Logger::defaultTags() const {
+  return Tags();
+}
 
-// logger that writes to standard out
-class StdOutLogger : public Logger {
+// Std stream worker.
+class StdWorker {
 public:
-  StdOutLogger() = delete;
-  StdOutLogger(const LoggingConfig& config)
-      : Logger(config),
-        levels(config.find("color") != config.end() && config.find("color")->second == "true"
-                   ? colored
-                   : uncolored) {
+  // Construct worker using its configuration.
+  explicit StdWorker(const LoggingConfig& config, std::ostream& stream)
+      : highlighted_(config.find("color") != config.end() && config.find("color")->second == "true"),
+        stream_(stream) {
   }
-  virtual void Log(const std::string& message, const LogLevel level) {
+
+  void operator()(const Message& message) const {
 #ifdef __ANDROID__
-    __android_log_print(android_levels.find(level)->second, "valhalla", "%s", message.c_str());
+    const auto& tags = message.tags();
+    __android_log_print(android_priority(tags), "valhalla", "%s\n",
+                        (tags.str() + ": " + message.serialize()).c_str());
 #else
-    Log(message, levels.find(level)->second);
-#endif
-  }
-  virtual void Log(const std::string& message, const std::string& custom_directive = " [TRACE] ") {
-#ifdef __ANDROID__
-    __android_log_print(ANDROID_LOG_INFO, "valhalla", "%s", message.c_str());
-#else
-    std::string output;
-    output.reserve(message.length() + 64);
-    output.append(TimeStamp());
-    output.append(custom_directive);
-    output.append(message);
-    output.push_back('\n');
-    // cout is thread safe, to avoid multiple threads interleaving on one line
-    // though, we make sure to only call the << operator once on std::cout
-    // otherwise the << operators from different threads could interleave
-    // obviously we dont care if flushes interleave
-    std::cout << output;
-    std::cout.flush();
+    std::stringstream ss;
+    ss << TimeStamp() << " ";
+    const auto& tags = message.tags();
+    if (highlighted_) {
+      for (const auto& tag : tags.tags()) {
+        std::string enclosed = "[" + tag.str() + "]";
+        ss << (tag.custom() ? enclosed : highlight(enclosed));
+      }
+    } else {
+      ss << message.tags().str();
+    }
+
+    if (!message.tags().empty())
+      ss << " ";
+    ss << message.serialize() << std::endl;
+    stream_ << ss.rdbuf() << std::flush;
 #endif
   }
 
 protected:
-  const std::unordered_map<LogLevel, std::string, EnumHasher> levels;
+  // Determines if special tags should be highlighted.
+  const bool highlighted_;
+  std::ostream& stream_;
 };
-bool std_out_logger_registered = RegisterLogger("std_out", [](const LoggingConfig& config) {
-  Logger* l = new StdOutLogger(config);
-  return l;
-});
 
-class StdErrLogger : public StdOutLogger {
-  using StdOutLogger::StdOutLogger;
-  virtual void Log(const std::string& message, const std::string& custom_directive = " [TRACE] ") {
-#ifdef __ANDROID__
-    __android_log_print(ANDROID_LOG_ERROR, "valhalla", "%s", message.c_str());
-#else
-    std::string output;
-    output.reserve(message.length() + 64);
-    output.append(TimeStamp());
-    output.append(custom_directive);
-    output.append(message);
-    output.push_back('\n');
-    std::cerr << output;
-    std::cerr.flush();
-#endif
+// Stdout worker.
+class StdOutWorker : public StdWorker {
+public:
+  // Construct worker using its configuration.
+  explicit StdOutWorker(const LoggingConfig& config) : StdWorker(config, std::cout) {
   }
 };
-bool std_err_logger_registered = RegisterLogger("std_err", [](const LoggingConfig& config) {
-  Logger* l = new StdErrLogger(config);
-  return l;
-});
 
-// TODO: add log rolling
-// logger that writes to file
-class FileLogger : public Logger {
+// Stderr worker.
+class StdErrWorker : public StdWorker {
 public:
-  FileLogger() = delete;
-  FileLogger(const LoggingConfig& config) : Logger(config) {
-    // grab the file name
+  // Construct worker using its configuration.
+  explicit StdErrWorker(const LoggingConfig& config) : StdWorker(config, std::cerr) {
+  }
+};
+
+// Text file worker.
+class TextFileWorker {
+public:
+  // Construct worker using its configuration.
+  explicit TextFileWorker(const LoggingConfig& config) {
+    // Grab the file name.
     auto name = config.find("file_name");
     if (name == config.end()) {
       throw std::runtime_error("No output file provided to file logger");
     }
     file_name = name->second;
 
-    // if we specify an interval
-    reopen_interval = std::chrono::seconds(300);
-    auto interval = config.find("reopen_interval");
-    if (interval != config.end()) {
-      try {
-        reopen_interval = std::chrono::seconds(std::stoul(interval->second));
-      } catch (...) {
-        throw std::runtime_error(interval->second + " is not a valid reopen interval");
-      }
-    }
+    // Open file.
+    open();
+  }
 
-    // crack the file open
-    ReOpen();
-  }
-  virtual void Log(const std::string& message, const LogLevel level) {
-    Log(message, uncolored.find(level)->second);
-  }
-  virtual void Log(const std::string& message, const std::string& custom_directive = " [TRACE] ") {
-    std::string output;
-    output.reserve(message.length() + 64);
-    output.append(TimeStamp());
-    output.append(custom_directive);
-    output.append(message);
-    output.push_back('\n');
-    lock.lock();
-    file << output;
-    file.flush();
-    lock.unlock();
-    ReOpen();
+  void operator()(const Message& message) {
+    const auto& tags = message.tags();
+    std::stringstream ss;
+    ss << TimeStamp() << " " << tags.str();
+    if (!tags.empty())
+      ss << " ";
+    ss << message.serialize() << std::endl;
+    file << ss.rdbuf() << std::flush;
   }
 
 protected:
-  void ReOpen() {
-    // TODO: use CLOCK_MONOTONIC_COARSE
-    // check if it should be closed and reopened
-    auto now = std::chrono::system_clock::now();
-    lock.lock();
-    if (now - last_reopen > reopen_interval) {
-      last_reopen = now;
-      try {
-        file.close();
-      } catch (...) {}
-      try {
-        file.open(file_name, std::ofstream::out | std::ofstream::app);
-        last_reopen = std::chrono::system_clock::now();
-      } catch (std::exception& e) {
-        try {
-          file.close();
-        } catch (...) {}
-        throw e;
-      }
-    }
-    lock.unlock();
+  // Open file to be written to.
+  void open() {
+    file.open(file_name, std::fstream::out | std::fstream::app);
+    if (!file.good())
+      throw std::runtime_error("Unable to open log file");
   }
+
+  // File name and stream to be written to.
   std::string file_name;
-  std::ofstream file;
-  std::chrono::seconds reopen_interval;
-  std::chrono::system_clock::time_point last_reopen;
+  std::fstream file;
 };
-bool file_logger_registered = RegisterLogger("file", [](const LoggingConfig& config) {
-  Logger* l = new FileLogger(config);
-  return l;
-});
 
 } // namespace logging
 
-// statically get a logger using the factory
-logging::Logger& logging::GetLogger(const LoggingConfig& config) {
-  static std::unique_ptr<Logger> singleton(GetFactory().Produce(config));
-  return *singleton;
-}
-
-// statically log manually
-void logging::Log(const std::string& message, const logging::LogLevel level) {
-  GetLogger().Log(message, level);
-}
-
-// statically log manually
-void logging::Log(const std::string& message, const std::string& custom_directive) {
-  GetLogger().Log(message, custom_directive);
+// statically get a logger
+logging::Logger& logging::GetLogger() {
+  static valhalla::midgard::logging::Logger
+      logger(valhalla::midgard::logging::Worker::create<valhalla::midgard::logging::StdOutWorker>(
+                 "stdout"),
+             valhalla::midgard::logging::Worker::create<valhalla::midgard::logging::StdErrWorker>(
+                 "stderr"),
+             valhalla::midgard::logging::Worker::create<valhalla::midgard::logging::TextFileWorker>(
+                 "file"));
+  return logger;
 }
 
 // statically configure logging
 void logging::Configure(const LoggingConfig& config) {
-  GetLogger(config);
+  GetLogger().configure(config);
 }
 
 } // namespace midgard