+++ CHANGELOG.md
@@ -29,6 +29,7 @@
    * ADDED: Allow using googletest in unit tests and convert all tests to it (old test.cc is completely removed). [#2128](https://github.com/valhalla/valhalla/pull/2128)
    * ADDED: Add guidance view capability. [#2209](https://github.com/valhalla/valhalla/pull/2209)
    * ADDED: Collect turn cost information as path is formed so that it can be seralized out for trace attributes or osrm flavored intersections. Also add shape_index to osrm intersections. [#2207](https://github.com/valhalla/valhalla/pull/2207)
+   * UPDATED: Logger. Asynchronous version added. [#2216](https://github.com/valhalla/valhalla/pull/2216)
 
 ## Release Date: 2019-11-21 Valhalla 3.0.9
 * **Bug Fix**